chore: update add issues to project action
The GitHub projects API has been replaced with ProjectsV2. This PR corrects the add issues to project action.

Also removes the external GH action and replaces with a simple GQL mutation for labelling

There should be no impact on any FE functionality - this affects Penny and I in terms of tracking new issues